moved singlepage and multipage tiff classes to their own modules started work on brute force ome-tiff output dims standardization
